requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
chromadb
langchain
langchain-openai
openai
requests
beautifulsoup4
lxml
pydantic
PyJWT[crypto]
